Update ChildDevPaper.Rmd
Made a temp fix to get the report to work, I'll come back to it and fix it later. I also increased the size of the bias_grid image so that the labels could be read. If that's not ok just go to line 461 and change the fig.width and fig.height to there previous settings. The document needed the load the kableExtra package so I added it. I also changed the default options for knitr to match the settings found in other reports.